Add template embeding via go:embed.
This change promts a significant change to how templates are being
handled. There is now a `go generate` to be done before building as
otherwise the application can't embed the required templates. I try to
be as efficient as I can with my little knowledge about how to handle
this.